@@ -1,56 +1,36 @@
 ﻿/*
  * Copyright ©
- * 2023
+ * 2025
  * Osella Ventures, LLC
  * All Rights Reserved
 */
 
-using NeuralNetworkLib;
-using TestSetupLib;
+using SetupLib;
 
-namespace NeuralNetworkMain
+namespace NeuralNetworkExec;
+
+public class NeuralNetworkMain
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        static void Main(string[] args)
+        try
         {
-            try
+            if (args.Length < 1)
             {
-                if (args.Length < 1)
-                {
-                    throw new ArgumentException("Usage: NeuralNetwork <path to the JSON test setup file>");
-                }
-
-                // Generate the Test Setup (Neural Network Structure and Training / Testing Samples)
-                var testSetup = TestSetupReader.GetTestSetup(args[0]);
-
-                // Create a Neural Network using the structure specified in the Test Setup
-                var neuralNetwork = new NeuralNetwork(
-                    testSetup.NbrInputs,
-                    testSetup.LayerConfigs,
-                    testSetup.Rnd,
-                    testSetup.UserDefinedFunctions);
-
-                // Train the Neural Network using the training samples generated by the Test Setup
-                neuralNetwork.Train(
-                    testSetup.Samples.TrainingInputs,
-                    testSetup.Samples.TrainingTargets,
-                    testSetup.NbrEpochs,
-                    testSetup.TrainingRate,
-                    testSetup.TrainingMomentum);
-
-                // Test the Neural Network using the testing samples generated by the Test Setup
-                neuralNetwork.Test(
-                    testSetup.Samples.TestingInputs,
-                    testSetup.Samples.TestingTargets);
-
-                // Summarize the Test Results
-                testSetup.UserDefinedFunctions.SummarizeTestResults();
+                throw new ArgumentException("Usage: NeuralNetwork <path to the JSON test setup file>");
             }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
-            }
+
+            // Read and Generate the Neural Network Setup (Neural Network Structure and Training / Testing Samples)
+            var setup = SetupReader.GetNeuralNetworkSetup(args[0]);
+
+            NeuralNetworkRunner.RunNeuralNetwork(setup);
+
+            Console.WriteLine("\nPress any key to exit...");
+            Console.ReadKey();
+        }
+        catch (Exception ex)
+        {
+            Console.WriteLine(ex.Message);
         }
     }
 }
