@@ -1,33 +1,32 @@
 ﻿/*
  * Copyright ©
- * 2023
+ * 2025
  * Osella Ventures, LLC
  * All Rights Reserved
 */
 
-namespace NeuralNetworkLib.ActivationFunctions
+namespace NeuralNetworkLib.ActivationFunctions;
+
+public class ParametricReLUActivationFunction : IActivationFunction
 {
-    public class ParametricReLUActivationFunction : IActivationFunction
+    private readonly double parameter;
+
+    public ParametricReLUActivationFunction(double parameter)
     {
-        private readonly double parameter;
+        this.parameter = parameter;
+    }
 
-        public ParametricReLUActivationFunction(double parameter)
-        {
-            this.parameter = parameter;
-        }
+    public double Compute(double x)
+    {
+        return x < 0.0 ? (parameter * x) : x;
+    }
 
-        public double Compute(double x)
-        {
-            return x < 0.0 ? (parameter * x) : x;
-        }
+    public double Derivative(double x, double fofx)
+    {
+        return x < 0.0 ? parameter : 1.0;
+    }
 
-        public double Derivative(double x, double fofx)
-        {
-            return x < 0.0 ? parameter : 1.0;
-        }
-
-        public void PostProcess(double[] ary)
-        {
-        }
+    public void PostProcess(double[] ary)
+    {
     }
 }
