@@ -1,35 +1,34 @@
 ﻿/*
  * Copyright ©
- * 2023
+ * 2025
  * Osella Ventures, LLC
  * All Rights Reserved
 */
 
-namespace NeuralNetworkLib.ActivationFunctions
+namespace NeuralNetworkLib.ActivationFunctions;
+
+public class TanhActivationFunction : IActivationFunction
 {
-    public class TanhActivationFunction : IActivationFunction
+    public double Compute(double x)
     {
-        public double Compute(double x)
+        if (x < -20.0)
         {
-            if (x < -20.0)
-            {
-                return -1.0;
-            }
-            else if (x > 20.0)
-            {
-                return 1.0;
-            }
-
-            return Math.Tanh(x);
+            return -1.0;
+        }
+        else if (x > 20.0)
+        {
+            return 1.0;
         }
 
-        public double Derivative(double x, double fofx)
-        {
-            return 1.0 - (fofx * fofx);
-        }
+        return Math.Tanh(x);
+    }
 
-        public void PostProcess(double[] ary)
-        {
-        }
+    public double Derivative(double x, double fofx)
+    {
+        return 1.0 - (fofx * fofx);
+    }
+
+    public void PostProcess(double[] ary)
+    {
     }
 }
