@@ -1,45 +1,44 @@
 ﻿/*
  * Copyright ©
- * 2023
+ * 2025
  * Osella Ventures, LLC
  * All Rights Reserved
 */
 
-namespace SamplesGeneratorLib
+namespace SamplesGeneratorLib;
+
+// Abstract base class for Samples Generators
+public abstract class SamplesGenerator
 {
-    // Abstract base class for Samples Generators
-    public abstract class SamplesGenerator
+    protected int nbrValuesPerRecord;   // number of inputs (defined or calculated) + nbrOutputs
+    protected double[][] records;       // number of records x nbrValuesPerRecord 
+
+    protected readonly int nbrOutputs;
+
+    protected SamplesGenerator(
+        int nbrOutputs)
     {
-        protected int nbrValuesPerRecord;   // number of inputs (defined or calculated) + nbrOutputs
-        protected double[][] records;       // number of records x nbrValuesPerRecord 
-
-        protected readonly int nbrOutputs;
-
-        protected SamplesGenerator(
-            int nbrOutputs)
+        if (nbrOutputs < 1)
         {
-            if (nbrOutputs < 1)
-            {
-                throw new InvalidOperationException();
-            }
-
-            nbrValuesPerRecord = 0;
-
-            this.nbrOutputs = nbrOutputs;
+            throw new InvalidOperationException();
         }
 
-        protected Samples GetSamples(
-            double trainingFraction,
-            bool normalizeInputs,
-            Random rnd = null)
-        {
-            return new Samples(
-                trainingFraction,
-                normalizeInputs,
-                records,
-                nbrValuesPerRecord,
-                nbrOutputs,
-                rnd);
-        }
+        nbrValuesPerRecord = 0;
+
+        this.nbrOutputs = nbrOutputs;
+    }
+
+    protected Samples GetSamples(
+        double trainingFraction,
+        bool normalizeInputs,
+        Random rnd = null)
+    {
+        return new Samples(
+            trainingFraction,
+            normalizeInputs,
+            records,
+            nbrValuesPerRecord,
+            nbrOutputs,
+            rnd);
     }
 }
